Make Didomi integration a bit more robust
- Introduce utility file for page scripts and use it
- Always use the object instead of the DOM element to check readiness.
The DOM element may show up before we can do
window.Didomi.getUserStatus, so it's best to refer to the object.
